+++ app/Controller/Admin/UserController.php
@@ -7,8 +7,11 @@ namespace App\Controller\Admin;
 use App\Controller\AbstractController;
 use App\Model\AdminPermission;
 use App\Model\AdminRole;
+use App\Model\Comment;
+use App\Model\Post;
 use App\Model\User;
 use App\Request\UserRequest;
+use App\Resource\NavResource;
 use App\Resource\UserResource;
 use Donjan\Casbin\Enforcer;
 use Hyperf\HttpServer\Annotation\Controller;
@@ -23,16 +26,40 @@ use Phper666\JWTAuth\JWT;
  * Class UserController
  * @package App\Controller\Admin
  * @Controller()
- * @Middlewares({
- *     @Middleware(JWTAuthMiddleware::class),
- *     @Middleware(PermissionMiddleware::class)
- * })
  */
 class UserController extends AbstractController
 {
     /**
      * @return \Psr\Http\Message\ResponseInterface
+     * @RequestMapping(path="nav", methods="get")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class)
+     * })
+     */
+    public function nav(JWT $JWT)
+    {
+        $user = $JWT->getParserData();
+        $role_id = Enforcer::getRolesForUser('roles_' . $user['id'])[0];
+        $permission = Enforcer::getPermissionsForUser('permission_' . $role_id);
+        $permissions = array();
+        foreach ($permission as $k => $v) {
+            //每一项权限
+            $permission_item = NavResource::make(AdminPermission::query()->where(['id'=> $v[3], 'is_menu'=> 1])->orderBy('sort', 'asc')->first());
+            $permissions[$k] = $permission_item;
+        }
+        $data = NavResource::collection(AdminPermission::query()->where(['p_id'=> 0, 'is_menu'=> 1])->orderBy('sort', 'asc')->get());
+        foreach ($data as $k => $v){
+            array_push($permissions, $data[$k]);
+        }
+        return $this->success($permissions);
+    }
+
+    /**
+     * @return \Psr\Http\Message\ResponseInterface
      * @RequestMapping(path="info", methods="get")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class)
+     * })
      */
     public function info(JWT $JWT)
     {
@@ -44,7 +71,7 @@ class UserController extends AbstractController
         foreach ($permission as $k => $v) {
             //每一项权限
             $permission_item = AdminPermission::query()->where('id', $v[3])->first();
-            $method = json_decode($permission_item['method']);
+            $method = $permission_item['method'];
             $methods = [
                 [
                     'action' => 'POST',
@@ -66,7 +93,7 @@ class UserController extends AbstractController
             ];
             $permission_method = array();
             foreach ($method as $k2 => $v2) {
-                if(in_array($methods[$k2]['action'], $method)){
+                if (in_array($methods[$k2]['action'], $method)) {
                     array_push($permission_method, $methods[$k2]['action']);
                 }
             }
@@ -132,7 +159,7 @@ class UserController extends AbstractController
                 ['ip', 'like', $ip],
                 ['email', 'like', $email],
             ])
-            ->paginate((int) $pageSize, ['*'], 'page', (int) $pageNo);
+            ->paginate((int)$pageSize, ['*'], 'page', (int)$pageNo);
         $users = $user->toArray();
 
         $data = [
@@ -148,6 +175,10 @@ class UserController extends AbstractController
     /**
      * @return \Psr\Http\Message\ResponseInterface
      * @RequestMapping(path="create", methods="post")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class),
+     *     @Middleware(PermissionMiddleware::class)
+     * })
      */
     public function create(UserRequest $request)
     {
@@ -155,29 +186,34 @@ class UserController extends AbstractController
         $data = $request->validated();
         $data['password'] = $this->passwordHash($data['password']);
         $flag = (new UserResource(User::query()->create($data)))->toResponse();
-        if($flag){
+        if ($flag) {
             return $this->success();
         }
         return $this->fail();
     }
 
     /**
+     * @param UserRequest $request
      * @param int $id
      * @return \Psr\Http\Message\ResponseInterface
      * @RequestMapping(path="update/{id}", methods="put")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class),
+     *     @Middleware(PermissionMiddleware::class)
+     * })
      */
     public function update(UserRequest $request, int $id)
     {
         // 验证
         $data = $request->validated();
         $data['password'] = $this->passwordHash($data['password']);
-        if(!Enforcer::hasRoleForUser('roles_' . $id, $data['user_role'])){
+        if (!Enforcer::hasRoleForUser('roles_' . $id, $data['user_role'])) {
             Enforcer::deleteRolesForUser('roles_' . $id);
             Enforcer::addRoleForUser('roles_' . $id, $data['user_role']);
         }
         unset($data['user_role']);
         $flag = User::query()->where('id', $id)->update($data);
-        if($flag){
+        if ($flag) {
             return $this->success();
         }
         return $this->fail();
@@ -187,6 +223,10 @@ class UserController extends AbstractController
      * @param int $id
      * @return \Psr\Http\Message\ResponseInterface
      * @RequestMapping(path="edit/{id}", methods="post")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class),
+     *     @Middleware(PermissionMiddleware::class)
+     * })
      */
     public function edit(int $id)
     {
@@ -197,11 +237,23 @@ class UserController extends AbstractController
      * @param int $id
      * @return \Psr\Http\Message\ResponseInterface
      * @RequestMapping(path="delete/{id}", methods="delete")
+     * @Middlewares({
+     *     @Middleware(JWTAuthMiddleware::class),
+     *     @Middleware(PermissionMiddleware::class)
+     * })
      */
     public function delete(int $id)
     {
+        //判断用户存在文章
+        if(Post::query()->where('author', $id)->first()){
+            return $this->fail([], '用户存在文章');
+        }
+        //判断用户存在评论
+        if(Comment::query()->where('user_id', $id)->first()){
+            return $this->fail([], '用户存在评论');
+        }
         $flag = User::query()->where('id', $id)->delete();
-        if($flag){
+        if ($flag) {
             return $this->success();
         }
         return $this->fail();
